adding _start to kernel_entry.asm
When running this command:

```
$ i386-elf-ld -o /rootfs/boot/kernel.bin -Ttext 0x1000 /code/kernel/kernel_entry.o /code/kernel/kernel.o --oformat binary
```
I got an error about the `_start` missing

```
i386-elf-ld: warning: cannot find entry symbol _start; defaulting to 0000000000001000
```
And the fix was to add it (see PR). I'm testing the build from scratch to see if there are any new issues that arise, but at least in basic testing this seemed ok to me! I'm very new at assembly / C so apologies in advance if I am missing something!