Finished implementing int_to_ascii
I've been following along with your tutorials, and I noticed that your int_to_ascii function still needs to reverse the contents of the string it populates. I implemented some code that reverses the string in place and tested it. Hopefully this is helpful, thank you so much for your awesome tutorials!